Implemented new isNodesetExpr and asNode methods on Expression, and getArg0AsNode and Arg0IsNodesetExpr on FunctionDef1Arg. This allows functions that just take a single node to get that node without creating an XNodeset or iterator if the expression is a simple one-step LocationPath (more complicated LocationPaths still need iterators, at least for the moment).  There will be more to come along these lines, trying to reduce the number of XObject derivatives that are are created during function evaluation.
Added sendNormalizedSAXcharacters to the FastStringBuffer, and
a param on DTM#dispatchCharactersEvents to enable normalization
of the characters.  This allows us to send normalized characters
without copying.  I think the sendNormalizedSAXcharacters may
need a little tuning.

Added tweak/hack to DTMDefaultBase which is a flag telling
if a namespace has been seen, and, if it has not, always return
immediately from getFirstNamespaceNode, which, as it turns
out, is called a lot in the course of copy operations.  We need to
come up with something better than this.  I suggested to Joe
that we use a bitset for namespaces and attributes that will
tell us if a given element has these nodes (I think it would help
for getFirstAttribute as well).


git-svn-id: https://svn.apache.org/repos/asf/xalan/java/branches/DTM_EXP@334881 13f79535-47bb-0310-9956-ffa450edef68

+++ src/org/apache/xml/dtm/ref/DTMDefaultBaseTraversers.java
@@ -1090,7 +1090,7 @@ public abstract class DTMDefaultBaseTraversers extends DTMDefaultBase
       for (current = (current & m_mask) - 1; current >= 0; current--)
       {
         int exptype = m_exptype[current];
-        int type = ExpandedNameTable.getType(exptype);
+        short type = ExpandedNameTable.getType(exptype);
 
         if (ATTRIBUTE_NODE == type || NAMESPACE_NODE == type
                 || isAncestor(subtreeRootIdent, current))
@@ -1120,7 +1120,7 @@ public abstract class DTMDefaultBaseTraversers extends DTMDefaultBase
       for (current = (current & m_mask) - 1; current >= 0; current--)
       {
         int exptype = m_exptype[current];
-        int type = ExpandedNameTable.getType(exptype);
+        short type = ExpandedNameTable.getType(exptype);
 
         if (exptype != extendedTypeID
                 || isAncestor(subtreeRootIdent, current))
@@ -1156,7 +1156,7 @@ public abstract class DTMDefaultBaseTraversers extends DTMDefaultBase
       for (current = (current & m_mask) - 1; current >= 0; current--)
       {
         int exptype = m_exptype[current];
-        int type = ExpandedNameTable.getType(exptype);
+        short type = ExpandedNameTable.getType(exptype);
 
         if (ATTRIBUTE_NODE == type || NAMESPACE_NODE == type)
           continue;
@@ -1185,7 +1185,7 @@ public abstract class DTMDefaultBaseTraversers extends DTMDefaultBase
       for (current = (current & m_mask) - 1; current >= 0; current--)
       {
         int exptype = m_exptype[current];
-        int type = ExpandedNameTable.getType(exptype);
+        short type = ExpandedNameTable.getType(exptype);
 
         if (exptype != extendedTypeID)
           continue;
